+++ out/extension.js
@@ -41,13 +41,11 @@ exports.deactivate = deactivate;
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 const vscode = __importStar(require("vscode"));
-const clockin_1 = __importDefault(require("./time/clockin"));
-const clockout_1 = __importDefault(require("./time/clockout"));
 const status_1 = __importDefault(require("./time/status"));
 const login_1 = __importDefault(require("./email/login"));
 const check_screen_time_1 = __importDefault(require("./screentime/check_screen_time"));
-const logout_1 = __importDefault(require("./email/logout"));
 const setGoal_1 = __importDefault(require("./time/setGoal"));
+const session_end_1 = __importDefault(require("./email/session_end"));
 // This method is called when extension is activated
 // extension is activated the very first time the command is executed
 function activate(context) {
@@ -55,17 +53,24 @@ function activate(context) {
     console.log('Congratulations, your extension "work-progress" is now active!');
     // Check how long the person has been working
     (0, check_screen_time_1.default)(context);
+    if (context.globalState.get("time_worked") !== undefined && context.globalState.get("time_worked") !== 0 && vscode.workspace.getConfiguration("work-progress").get("session", false)) {
+        console.log("sending email for time worked: " + context.globalState.get("time_worked"));
+        (0, session_end_1.default)(context, parseInt(context.globalState.get("time_worked") || "0") / 60);
+        context.globalState.update("time_worked", "0");
+    }
     const disposable = vscode.commands.registerCommand('work-progress.helloWorld', () => {
         // The code you place here will be executed every time your command is executed
         // Display a message box to the user
         vscode.window.showInformationMessage('Work Progress has been loaded correctly!');
     });
-    context.subscriptions.push(disposable, vscode.commands.registerCommand('work-progress.setGoal', () => (0, setGoal_1.default)(context)), vscode.commands.registerCommand("work-progress.login", () => (0, login_1.default)(context)), vscode.commands.registerCommand("work-progress.logout", () => (0, logout_1.default)(context)), vscode.commands.registerCommand('work-progress.clockIn', () => (0, clockin_1.default)(context)), vscode.commands.registerCommand('work-progress.clockOut', () => (0, clockout_1.default)(context)), vscode.commands.registerCommand('work-progress.status', () => (0, status_1.default)(context)));
+    context.subscriptions.push(disposable, vscode.commands.registerCommand('work-progress.setGoal', () => (0, setGoal_1.default)(context)), vscode.commands.registerCommand("work-progress.login", () => (0, login_1.default)(context)), vscode.commands.registerCommand('work-progress.status', () => (0, status_1.default)(context)));
 }
 // This method is called when your extension is deactivated
 function deactivate(context) {
+    // sessionEnd(context, context.globalState.get("time_worked") || 0);
+    // Clear the global state
+    console.log("time worked: " + context.globalState.get("time_worked"));
     // This log helps with cathing errors at the shutdown
-    console.log('Work Progress has been deactivated!');
-    (0, clockout_1.default)(context);
+    ;
 }
 //# sourceMappingURL=extension.js.map