+++ src/time/status.ts
@@ -1,11 +1,10 @@
-import { endTimer } from "./time_count";
+
 import * as vscode from "vscode";
 
 export default function status(context: vscode.ExtensionContext) {
-    const start_time = parseInt(context.globalState.get("start_time") || "0");
-    const status = endTimer(start_time);
-    if (start_time === 0) {
-        vscode.window.showInformationMessage("You have not clocked in yet!");
+    const status = parseInt(context.globalState.get("time_worked") || "0");
+    if (status === 0) {
+        vscode.window.showInformationMessage("error occured");
     }
     else {
         vscode.window.showInformationMessage("You have worked " + status + " minutes! Keep going!");