+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log
 
 All notable changes to the "work-progress" extension will be documented in this file.
+ 
+- Sign in using api created!
+- Website!!!
+## 1.0.1:
 
-- Release in the visual studio code marketplace.
-- Added message when clocked in
-- NEW set goal command, now you can set your working goal
-- Fixed bugs
+- Minor changes, made email sending api more save
+
+
+## 1.0.2 (NOT in vscode marketplace yet)
+
+- Emails sent after each session (have to enable, sorry)
+- Time measured without having to enter commands
+- API is now manageble from settings